requirements.txt
Flask
quick-mailer
python-dotenv